ModelDiff: Calculate the difference between two models
ModelDiff is a class that will create separate dicts of all added, removed,
changed and unchanged Users when passed a dict of Users from the source;
a dict of Users from the target; and a config dict specifying which
fields in the source are important, and regex patterns of all the groups
from the source that are important.

The "changed_users" property contains Users that have been merged from
the two - any field specified in the config will be taken from the
source, and any field not specified will be taken from the target.

Similarly, it merges the list of groups by taking from the source
any groups whose names match any of the regexes provided, and takes all
groups that *don't* match any of the regexes from the target.