Remove the 'module' field from config before parsing
This field is mandatory, and used to find out which module to load the
config with, but there's no benefit to passing it into the module's
configuration.

Removing it means that we don't have to specify that we expect a
'module' field in the config every time we write a new module.