target_suitecrm: Fix creating Users using the wrong type for groups
The default type for groups is a tuple, not a list. This is important,
because a List is not a hashable and orderable type (i.e. if given a
list of two lists, how would you decide which one goes first). This is
important for Users to also be hashable and orderable, so that we can
compare Users.