target_suitecrm: Stop trying to set fullname as a unique thing
In our model (and freeipa), a user's full name will default to a
concatenation of forename and surname, but can be set to something
distinct.
In SuiteCRM, a user's full name is *always* a concatenation of forename.
e.g. if passed the full name "John P. Smith", it will store forename
"John" and surname "P. Smith".

Thus, we should stop trying to push "name" and "full_name" to suitecrm,
and we should stop expecting fullname to be identical between two Users
when suitecrm is involved.